added code of junit and Exception handler

+++ CustomerReward/src/main/java/com/reward/demo/service/RewardsService.java
@@ -16,8 +16,8 @@ public class RewardsService {
     @Autowired
     private TransactionRepository transactionRepository;
 
-    public int calculatePoints(double amount) {
-        int points = 0;
+    public double calculatePoints(double amount) {
+        double points = 0;
         if (amount > 100) {
             points += (amount - 100) * 2;
             amount = 100;
@@ -28,16 +28,16 @@ public class RewardsService {
         return points;
     }
 
-    public int getMonthlyPoints(String customerId, int month, int year) {
+    public double getMonthlyPoints(String customerId, int month, int year) {
         LocalDate startDate = LocalDate.of(year, month, 1);
         LocalDate endDate = startDate.withDayOfMonth(startDate.lengthOfMonth());
         List<Transaction> transactions = transactionRepository.findByCustomerIdAndDateBetween(customerId, startDate, endDate);
-        return transactions.stream().mapToInt(t -> calculatePoints(t.getAmount())).sum();
+        return transactions.stream().mapToDouble(t -> calculatePoints(t.getAmount())).sum();
     }
 
-    public int getTotalPoints(String customerId) {
+    public double getTotalPoints(String customerId) {
         List<Transaction> transactions = transactionRepository.findByCustomerIdAndDateBetween(customerId, LocalDate.now().minusMonths(3), LocalDate.now());
-        return transactions.stream().mapToInt(t -> calculatePoints(t.getAmount())).sum();
+        return transactions.stream().mapToDouble(t -> calculatePoints(t.getAmount())).sum();
     }
 }
